backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
alembic
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5
pulp
